requirements.txt
pycryptodome
allure-pytest==2.9.45
allure-python-commons==2.9.45

pytest==7.1.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-parallel==0.1.1
pytest-rerunfailures==10.2
pytest-xdist==2.5.0

requests==2.31.0
attrs>=21.4.0,<23.0.0
jinja2==3.1.2

Levenshtein==0.23.0
selenium==4.1.5
PyYAML==6.0.1
openpyxl==3.0.10

fastapi==0.104.1
uvicorn==0.24.0.post1
gradio==3.50.2
